requirements.txt
requests~=2.28.1
pandas~=1.4.3
PyMySQL~=1.0.2
SQLAlchemy~=1.4.40
azure.storage.blob